+++ module-admin-bemyplan/build.gradle
@@ -7,6 +7,7 @@ sourceSets {
 }
 
 dependencies {
+    implementation project(':module-domain-bemyplan')
     implementation platform('com.amazonaws:aws-java-sdk-bom:1.11.1000')
     implementation 'com.amazonaws:aws-java-sdk-s3'
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
