+++ module-admin-bemyplan/src/main/java/com/deploy/bemyplan/plan/creator/CreatorDto.java
@@ -1,0 +1,8 @@
+package com.deploy.bemyplan.plan.creator;
+
+import lombok.Getter;
+
+@Getter
+public class CreatorDto {
+    private String name;
+}
