+++ module-admin-bemyplan/src/main/java/com/deploy/bemyplan/plan/preview/PreviewDto.java
@@ -1,0 +1,9 @@
+package com.deploy.bemyplan.plan.preview;
+
+import lombok.Getter;
+
+@Getter
+public class PreviewDto {
+    private String description;
+    private String image;
+}
